% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_confidential_mode.R
\name{is_confidential_mode}
\alias{is_confidential_mode}
\title{Check whether confidential mode is active}
\usage{
is_confidential_mode()
}
\value{
Logical scalar.
}
\description{
Check whether confidential mode is active
}
