% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_fingerprint.R
\name{is_fingerprint}
\alias{is_fingerprint}
\title{Test whether an object is a confideR fingerprint}
\usage{
is_fingerprint(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Test whether an object is a confideR fingerprint
}
